added confirmation message before deletion in all the lists

+++ src/components/ListDeliverablesAPI.js
@@ -95,14 +95,18 @@ export const ListDeliverablesAPI = () => {
         setSearch(e.target.value);
     }
 
-    const deleteDeliverable = (deliverableID) => {
+    const deleteDeliverable = (deliverableID, house, date) => {
 
         console.log(deliverableID)
+        var confirmationResult = window.confirm(`Are you sure you want to delete the record of ${house} on ${date}?`)
+        if (confirmationResult)     //if confirmationResult===true
+        {
 
-        axios.delete(`http://localhost:4000/deliverables/` + deliverableID).then(res => {
-            console.log(res)
-            alert("Record deleted successfully!")
-        })
+            axios.delete(`http://localhost:4000/deliverables/` + deliverableID).then(res => {
+                console.log(res)
+                alert("Record deleted successfully!")
+            })
+        }
     }
 
     return (
@@ -142,7 +146,7 @@ export const ListDeliverablesAPI = () => {
                                         <td>{deliverables.isPickup ? <i className="bi bi-check-lg"></i> : <i className="bi bi-x-lg"></i>}</td>
                                         {localStorage.getItem('roleName') === 'Society Member' ?
                                             <td>
-                                                <Link to="/listdeliverables" className="btn btn-sm btn-danger mx-2" onClick={() => { deleteDeliverable(deliverables._id) }}><i className="bi bi-trash"></i></Link>
+                                                <Link to="/listdeliverables" className="btn btn-sm btn-danger mx-2" onClick={() => { deleteDeliverable(deliverables._id, deliverables.house.houseTitle, deliverables.date) }}><i className="bi bi-trash"></i></Link>
                                                 <Link to={`/update/${deliverables._id}`} className="btn btn-sm btn-primary my-1" value={deliverables._id}><i className="bi bi-pencil"></i></Link>
                                             </td> : ""}
                                     </tr>
@@ -162,7 +166,7 @@ export const ListDeliverablesAPI = () => {
                                             <td>{deliverables.ispickup ? <i className="bi bi-check-lg"></i> : <i className="bi bi-x-lg"></i>}</td>
                                             {localStorage.getItem('roleName') === 'Society Member' ?
                                                 <td>
-                                                    <Link to="/listdeliverables" className="btn btn-sm btn-danger mx-2" onClick={() => { deleteDeliverable(deliverables._id) }}><i className="bi bi-trash"></i></Link>
+                                                    <Link to="/listdeliverables" className="btn btn-sm btn-danger mx-2" onClick={() => { deleteDeliverable(deliverables._id, deliverables.house.houseTitle, deliverables.date) }}><i className="bi bi-trash"></i></Link>
                                                     <Link to={`/update/${deliverables._id}`} className="btn btn-sm btn-primary my-1" value={deliverables._id}><i className="bi bi-pencil"></i></Link>
                                                 </td> : ""}
                                         </tr>
@@ -175,14 +179,14 @@ export const ListDeliverablesAPI = () => {
                                         console.log("sorted member : " + sortedDeliverable)
                                         return (
                                             <tr key={sortedDeliverable._id}>
-                                                {/* <th scope="row">{member.user.firstName}</th> */}
+
                                                 <th scope="row">{counter}</th>
                                                 <td>{sortedDeliverable.house.houseTitle}</td>
                                                 <td>{sortedDeliverable.date}</td>
                                                 <td>{sortedDeliverable.ispickup ? <i className="bi bi-check-lg"></i> : <i className="bi bi-x-lg"></i>}</td>
                                                 {localStorage.getItem('roleName') === 'Society Member' ?
                                                     <td>
-                                                        <Link to="/listdeliverables" className="btn btn-sm btn-danger mx-2" onClick={() => { deleteDeliverable(sortedDeliverable._id) }}><i className="bi bi-trash"></i></Link>
+                                                        <Link to="/listdeliverables" className="btn btn-sm btn-danger mx-2" onClick={() => { deleteDeliverable(sortedDeliverable._id, deliverables.house.houseTitle, deliverables.date) }}><i className="bi bi-trash"></i></Link>
                                                         <Link to={`/update/${sortedDeliverable._id}`} className="btn btn-sm btn-primary my-1" value={sortedDeliverable._id}><i className="bi bi-pencil"></i></Link>
                                                     </td> : ""}
                                             </tr>
